Program a Keylogger
Keylogger is a program that can be implemented into a user's computer background to record what the user types. The program can be used for malicious data collection or hacking (e-mails, password etc.) A keylogger file can be sent straight to a hacker by email or stored on a computer.